broken build, cast long to int... :(
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/freenet@16830 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/node/Node.java
@@ -1007,8 +1007,8 @@ public class Node implements TimeSkewDetectorCallback {
 							e.printStackTrace();
 						}
 						//Perhaps a bit hackish...? Seems like this should be near it's definition in NodeStats.
-						nodeStats.avgStoreLocation.changeMaxReports(maxStoreKeys);
-						nodeStats.avgCacheLocation.changeMaxReports(maxCacheKeys);
+						nodeStats.avgStoreLocation.changeMaxReports((int)maxStoreKeys);
+						nodeStats.avgCacheLocation.changeMaxReports((int)maxCacheKeys);
 					}
 		});
